Added dgraph-js-http and put some things to work
User sign in with Github is workin. But has a issue to solve about UID mapping.
User login too. But just login and ipdate da page. (bugs to be fixed).
Added helpers to run queries and run mutations.
Create post is working.
Add a comment is working but has the issue with UID mapping.